Factored out components into smaller pieces
